requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
Werkzeug==3.1.3
